requirements.txt
Django==4.1.3
django-cache-url==3.4.2
django-debug-toolbar==3.2
django-livereload-server==0.4
django-modeltranslation==0.18.6
django-phonenumber-field==5.0.0
environs==9.3.0
Pillow==9.2.0